requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1

# AI / Embeddings
openai==1.30.1
tiktoken==0.7.0

# Vector Store
faiss-cpu==1.8.0
numpy==1.26.4

# PDF Processing
pypdf==4.2.0
langchain-text-splitters==0.2.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
tenacity==8.3.0
structlog==24.2.0
redis==5.0.4
celery[redis]==5.4.0
